[mediapipe-task-text] Upgrade native_assets_cli
