+++ packages/mediapipe-task-text/pubspec.yaml
@@ -24,11 +24,11 @@ dependencies:
   http: ^1.1.0
   logging: ^1.2.0
   mediapipe_core: ^0.0.1
-  native_assets_cli: ^0.3.0
-  native_toolchain_c: ^0.3.0
+  native_assets_cli: ^0.5.4
+  native_toolchain_c: ^0.4.1
   path: ^1.8.3
 
 dev_dependencies:
-  ffigen: ^9.0.1
+  ffigen: ^12.0.0
   flutter_lints: ^2.0.0
   test: ^1.21.0
